Fold ProjectUpdater into UpdateCommand
The build-project update logic lived in a separate static ProjectUpdater
helper, splitting the update command's behaviour across two files for no
reason. Move it into UpdateCommand as private methods so the command owns
its own logic, and delete ProjectUpdater.

Also drops the stale "Nuke" name (UpdateNukeCommonPackage ->
UpdateFalloutCommonPackage). No behaviour change.
