+++ src/Fallout.Cli/Commands/UpdateCommand.cs
@@ -2,11 +2,13 @@ using System.IO;
 using System.Threading.Tasks;
 using System.Linq;
 using System.Text.Json.Nodes;
+using NuGet.Versioning;
 using Fallout.Cli.Prompts;
 using Fallout.Common;
 using Fallout.Common.Execution;
 using Fallout.Common.IO;
 using Fallout.Solutions;
+using Fallout.Common.Tooling;
 using Fallout.Common.Tools.DotNet;
 using Fallout.Common.Utilities;
 using static Fallout.Common.Constants;
@@ -63,8 +65,65 @@ internal sealed class UpdateCommand : IFalloutCommand
     {
         var configuration = _configuration.Read(buildScript, evaluate: true);
         var projectFile = configuration[ConfigurationReader.BuildProjectFileKey];
+
         ProjectModelTasks.Initialize();
-        await ProjectUpdater.UpdateAsync(projectFile);
+        var buildProject = ProjectModelTasks.ParseProject(projectFile).NotNull();
+
+        UpdateTargetFramework(buildProject);
+        var previousPackageVersion = await UpdateFalloutCommonPackageAsync(buildProject);
+
+        if (previousPackageVersion.MinVersion >= NuGetVersion.Parse("0.23.5"))
+            RemoveLegacyFileIncludes(buildProject);
+
+        buildProject.Save();
+    }
+
+    private static void UpdateTargetFramework(Microsoft.Build.Evaluation.Project buildProject)
+    {
+        buildProject.SetProperty("TargetFramework", "net8.0");
+    }
+
+    private static async Task<FloatRange> UpdateFalloutCommonPackageAsync(Microsoft.Build.Evaluation.Project buildProject)
+    {
+        var packageItem = buildProject.Items.SingleOrDefault(x => x.EvaluatedInclude == FalloutCommonPackageId).NotNull();
+        var previousPackageVersion = FloatRange.Parse(packageItem.GetMetadataValue("Version"));
+
+        var latestPackageVersion = await NuGetVersionResolver.GetLatestVersion(FalloutCommonPackageId, includePrereleases: false);
+        if (!previousPackageVersion.Satisfies(NuGetVersion.Parse(latestPackageVersion)))
+            packageItem.SetMetadataValue("Version", latestPackageVersion);
+
+        return previousPackageVersion;
+    }
+
+    private static void RemoveLegacyFileIncludes(Microsoft.Build.Evaluation.Project buildProject)
+    {
+        var legacyIncludes =
+            new[]
+            {
+                "csproj.DotSettings",
+                "build.ps1",
+                "build.sh",
+                ".nuke",
+                "global.json",
+                "nuget.config",
+                "azure-pipelines.yml",
+                "Jenkinsfile",
+                "appveyor.yml",
+                ".travis.yml",
+                "GitVersion.yml"
+            };
+
+        buildProject.Xml.Items
+            .Where(x => x.ItemType == "None").ToList()
+            .Where(x => x.Include.ContainsAnyOrdinalIgnoreCase(legacyIncludes) ||
+                        x.Remove.ContainsAnyOrdinalIgnoreCase(legacyIncludes)).ToList()
+            .ForEach(x =>
+            {
+                var itemGroupElement = x.Parent;
+                itemGroupElement.RemoveChild(x);
+                if (itemGroupElement.Children.Count == 0)
+                    itemGroupElement.Parent.RemoveChild(itemGroupElement);
+            });
     }
 
     private void UpdateConfigurationFile(AbsolutePath rootDirectory)
